# config: Increase thin provisioning threshold