Collecting extend stats shows that extend takes between 2.2 to 6.2
seconds, with average of 3.7 seconds. With the default thresholds:

[irs]
volume_utilization_chunk_mb = 1024
volume_utilization_percent = 50

This means that we extend the volume when free space is 512 MiB. Writing
more than 512 MiB in 3.7 seconds (138.4 MiB/s) will cause the VM to
pause with ENOSPC.

This configuration was too low 10 years ago, and we need to update it
for modern storage. Update the values to allow 4 times faster writes
before we pause with ENOSPC.

With the new configuration:

[irs]
volume_utilization_chunk_mb = 2560
volume_utilization_percent = 20

We extend the volume when free space is 2048 MiB.

Testing with old and new configuration show that we can cope now with 4x
times faster write rate before we VMs pause during extend.

Before:

write rate  extends   pauses
----------------------------
 75 MiB/s        50        0
100 MiB/s        50        4
125 MiB/s        50        4
150 MiB/s        53       24

After:

write rate  extends   pauses
----------------------------
200 MiB/s        20        0
250 MiB/s        20        0
300 MiB/s        20        0
350 MiB/s        21        0
400 MiB/s        20        1
450 MiB/s        20        2
500 MiB/s        22        7
550 MiB/s        23        7

The downside of this change is allocating more space in the storage
domain. New empty disk will consume 2.5 GiB instead of 1 GiB.

Bug-Url: https://bugzilla.redhat.com/2051997
Signed-off-by: Nir Soffer <nsoffer@redhat.com>