@nirs 2022-02-22
> can you elaborate a little bit of what are the factors affecting the time it takes to extend? I mean besides the actual storage speed. Are there any waits in the process, communication with SPM, etc?

The time include lot of waiting since we use polling.

- libvirt event thread get a block threshold event, and mark the drive 
  for extension
- The periodic watermark monitor checks vms every 2 seconds. When it find
  that a drive needs extension, it sends a request to the spm by writing
  to the storage mailbox
- The SPM check storage every 2 seconds. When it find the request, it
  run the extend using the spm mailbox thread pool
- The host is polling its mailbox for replies every 2 seconds. When it detects
  the reply, it complete the extend on the host side and resume the vm if
  needed.

We cannot optimize much the storage mailbox since checking the mailbox requires
reading from storage, and many hosts may check the mailbox at the same time.
Maybe we can check every 1 second instead of 2.

We can optimize the block threshold event handling - it should really post an
event that will wake up the watermark monitor immediately, and start the extend
flow. This will save 0-2 seconds from the total time. This requires rewrite
of the watermark monitor and separating it from the periodic executor which
is a change I wanted to do for long time.