@nirs 2022-02-22
> linter says that `from vdsm.config import config` in `blockvolume_test` is unsed now and fails

flake8 is correct, fixed in current version.