@nirs 2022-02-22
> > Measuring write throughput should be done in the guest, there is big difference
> > between guest performance and host performance. Measuring is complected and it
> > is impossible to predict how fast the disk will grow on a system.
> 
> yes, but isn't it just that it's always somewhat slower in guest, never faster? In that case it is still useful as then we can measure easily on the host and estimate for the worst case of near-host performance in guest

I don't see how it can be faster in the guest. But it is not possible to measure the host
and assume that the measurement is correct for the future, because the load on the storage,
the network and the host at the time of the measurement can be different at the time the
guest use the storage.