@mykaul 2022-02-22
> 4x is a big improvement. The extra overhead is small enough, 1.5GB extra for today's SD sizes is worth it.
> 
> Still, with a bit over 6 seconds it means we will still eventually pause with >300MiB/s writes even in quiet conditions. Is it high enough? Should we go for 500 increasing the overhead? OTOH it's really not a big deal to pause occasionally, we just want to avoid frequent pauses every time e.g. etcd is extended.
> 
> @mykaul FYI

Looks good - time to change the defaults indeed.